Add shortcut ability for logic operators
In most languages (and now also in TomFrost's JEXL), logic operators
implement shortcuts: For example when evaluating `(A && B)`, the
expression `B` will not be evaluated when `A` is `False`. Similarly, in
`A || B`, the right operand will be ignored if `A` evaluates to `True`.

This extends our operators to support shortcut / lazy evaluation.
Operators now have a flag `evaluate_lazy` to denote this, which can be
used to implement custom operators as well.